Bug 877140 - Fix type of constructor argument for duration (r=mfinkle)

+++ mobile/android/base/animation/PropertyAnimator.java
@@ -53,11 +53,11 @@ public class PropertyAnimator implements Runnable {
     private FramePoster mFramePoster;
     private boolean mUseHardwareLayer;
 
-    public PropertyAnimator(int duration) {
+    public PropertyAnimator(long duration) {
         this(duration, new DecelerateInterpolator());
     }
 
-    public PropertyAnimator(int duration, Interpolator interpolator) {
+    public PropertyAnimator(long duration, Interpolator interpolator) {
         mDuration = duration;
         mDurationReciprocal = 1.0f / (float) mDuration;
         mInterpolator = interpolator;
